Eliminate BLAS dependency for ALL crates (#207)
* Integrate ndarray-linalg-rs into linfa-clustering

* Restructure linfa-datasets to fix warnings

* Add support for linfa-ica

* Integrate into PLS (bugs pending)

* Integration with linfa-reduction complete

* Fix PLS CCA test

* Use invh instead of invc in elasticnet since the input might not be positive definite

* Integrate into linfa-preprocessing

* Integrate into linfa-linear

* Integrate into linfa-elasticnet

* Optimize gaussian mixture solve_triangular call

* Fix example Cargo invocations to not use blas features

* Remove BLAS from CI steps and add BLAS test job

* Enable most CI steps even in draft PRs

* Fix test-blas workflow name

* Eliminate BLAS from linfa-logistic

* Fix more examples invocations in READMEs

* Add non_exhaustive to modified error types so the blas feature is additive

* Change linfa-reduction to use truncated SVD and LOBPCG

* Switch to linfa-linalg

* Fix system BLAS to static BLAS in CI

* Centralized README docs on BLAS feature
