+++ .github/workflows/benching.yml
@@ -6,10 +6,6 @@ jobs:
   testing:
     name: benching
     runs-on: ubuntu-18.04
-    if: github.event.pull_request.draft == false
-    container:
-      image: rustmath/mkl-rust:1.43.0
-      options: --security-opt seccomp=unconfined
 
     steps:
       - name: Checkout sources
@@ -26,4 +22,4 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: bench
-          args: iai --all --features intel-mkl-system
+          args: iai --all
