+++ .github/workflows/checking.yml
@@ -6,7 +6,6 @@ jobs:
   check:
     name: check-${{ matrix.toolchain }}-${{ matrix.os }}
     runs-on: ${{ matrix.os }}
-    if: github.event.pull_request.draft == false
     strategy:
       fail-fast: false
       matrix:
