+++ .github/workflows/codequality.yml
@@ -7,7 +7,6 @@ jobs:
   codequality:
     name: codequality
     runs-on: ubuntu-latest
-    if: github.event.pull_request.draft == false
     strategy:
       matrix:
         toolchain:
@@ -42,9 +41,6 @@ jobs:
     name: coverage
     runs-on: ubuntu-18.04
     if: github.event.pull_request.draft == false
-    container:
-      image: rustmath/mkl-rust:1.43.0
-      options: --security-opt seccomp=unconfined
 
     steps:
       - name: Checkout sources
@@ -74,7 +70,7 @@ jobs:
 
       - name: Generate code coverage
         run: |
-          cargo tarpaulin --verbose --features intel-mkl-system --timeout 120 --out Xml --all --release
+          cargo tarpaulin --verbose --timeout 120 --out Xml --all --release
       - name: Upload to codecov.io
         uses: codecov/codecov-action@v1
         with:
