+++ .github/workflows/testing.yml
@@ -6,7 +6,6 @@ jobs:
   testing:
     name: testing-${{ matrix.toolchain }}-${{ matrix.os }}
     runs-on: ${{ matrix.os }}
-    if: github.event.pull_request.draft == false
     strategy:
       fail-fast: false
       matrix:
@@ -28,11 +27,38 @@ jobs:
           toolchain: ${{ matrix.toolchain }}
           override: true
 
-      - name: Log active toolchain
-        run: rustup show
-
-      - name: Run cargo test in release mode
+      - name: Run cargo test
         uses: actions-rs/cargo@v1
         with:
           command: test
-          args: --all --release --features intel-mkl-static
+          args: --release --workspace
+
+  testing-blas:
+    name: testing-with-BLAS-${{ matrix.toolchain }}-${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        toolchain:
+          - 1.54.0
+          - stable
+        os:
+          - ubuntu-18.04
+          - windows-2019
+
+    steps:
+      - name: Checkout sources
+        uses: actions/checkout@v2
+
+      - name: Install toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: ${{ matrix.toolchain }}
+          override: true
+
+      - name: Run cargo test with BLAS enabled
+        uses: actions-rs/cargo@v1
+        with:
+          command: test
+          args: --release --workspace --features intel-mkl-static,linfa-clustering/blas,linfa-ica/blas,linfa-reduction/blas,linfa-linear/blas,linfa-preprocessing/blas,linfa-pls/blas,linfa-elasticnet/blas
