+++ README.md
@@ -53,7 +53,7 @@ If this strikes a chord with you, please take a look at the [roadmap](https://gi
 
 ## BLAS/Lapack backend
 
-At the moment you can choose between the following BLAS/LAPACK backends: `openblas`, `netblas` or `intel-mkl`
+Some algorithm crates need to use an external library for linear algebra routines. By default, we use a pure-Rust implementation. However, you can also choose an external BLAS/LAPACK backend library instead, by enabling the `blas` feature and a feature corresponding to your BLAS backend. Currently you can choose between the following BLAS/LAPACK backends: `openblas`, `netblas` or `intel-mkl`.
 
 |Backend  | Linux | Windows | macOS |
 |:--------|:-----:|:-------:|:-----:|
@@ -61,11 +61,9 @@ At the moment you can choose between the following BLAS/LAPACK backends: `openbl
 |Netlib   |✔️      |-        |-      |
 |Intel MKL|✔️      |✔️        |✔️      |
 
-For example if you want to use the system IntelMKL library for the PCA example, then pass the corresponding feature:
-```
-cd linfa-reduction && cargo run --release --example pca --features linfa/intel-mkl-system
-```
-This selects the `intel-mkl` system library as BLAS/LAPACK backend. On the other hand if you want to compile the library and link it with the generated artifacts, pass `intel-mkl-static`.
+Each BLAS backend has two features available. The feature allows you to choose between linking the BLAS library in your system or statically building the library. For example, the features for the `intel-mkl` backend are `intel-mkl-static` and `intel-mkl-system`.
+
+An example set of Cargo flags for enabling the Intel MKL backend on an algorithm crate is `--features blas,linfa/intel-mkl-system`. Note that the BLAS backend features are defined on the `linfa` crate, and should only be specified for the final executable.
 
 # License
 Dual-licensed to be compatible with the Rust project.
