+++ algorithms/linfa-ica/Cargo.toml
@@ -14,7 +14,8 @@ categories = ["algorithms", "mathematics", "science"]
 
 [features]
 default = []
-serde = ["serde_crate", "ndarray/serde"]                                                     
+blas = ["ndarray-linalg", "linfa/ndarray-linalg"]
+serde = ["serde_crate", "ndarray/serde"]
 
 [dependencies.serde_crate]
 package = "serde"
@@ -25,14 +26,15 @@ features = ["std", "derive"]
 
 [dependencies]
 ndarray = { version = "0.15" }
-ndarray-linalg = "0.14"
+linfa-linalg = { version = "0.1", default-features = false }
+ndarray-linalg = { version = "0.14", optional = true }
 ndarray-rand = "0.14"
 ndarray-stats = "0.5"
 num-traits = "0.2"
 rand_xoshiro = "0.6"
 thiserror = "1.0"
 
-linfa = { version = "0.5.0", path = "../..", features = ["ndarray-linalg"] }
+linfa = { version = "0.5.0", path = "../.." }
 
 [dev-dependencies]
 ndarray-npy = { version = "0.8", default-features = false }
