+++ algorithms/linfa-logistic/Cargo.toml
@@ -14,11 +14,10 @@ keywords = ["machine-learning", "linfa", "ai", "ml", "linear"]
 categories = ["algorithms", "mathematics", "science"]
 
 [dependencies]
-ndarray = { version = "0.15", features = ["approx", "blas"] }
-ndarray-linalg = "0.14"
+ndarray = { version = "0.15", features = ["approx"] }
 ndarray-stats = "0.5.0"
 num-traits = "0.2"
-argmin = { version = "0.4.6", features = ["ndarrayl"] }
+argmin = { version = "0.4.6", features = ["ndarray", "ndarray-rand"] }
 serde = "1.0"
 thiserror = "1.0"
