+++ algorithms/linfa-logistic/src/float.rs
@@ -1,7 +1,6 @@
 use crate::argmin_param::ArgminParam;
 use argmin::prelude::{ArgminFloat, ArgminMul};
 use ndarray::{Dimension, Ix1, Ix2, NdFloat};
-use ndarray_linalg::Lapack;
 use num_traits::FromPrimitive;
 
 /// A Float trait that captures the requirements we need for the various
@@ -9,7 +8,6 @@ use num_traits::FromPrimitive;
 pub trait Float:
     ArgminFloat
     + NdFloat
-    + Lapack
     + Default
     + Clone
     + FromPrimitive
