+++ algorithms/linfa-pls/README.md
@@ -14,11 +14,15 @@
 
 ## Examples
 
-There is an usage example in the `examples/` directory. The example uses a BLAS backend, to run it and use the `intel-mkl` library do:
+There is an usage example in the `examples/` directory. To run it, use:
 
 ```bash
-$ cargo run --example pls_regression --features linfa/intel-mkl-system
+$ cargo run --example pls_regression
 ```
+
+## BLAS/Lapack backend
+
+See [this section](../../README.md#blaslapack-backend) to enable an external BLAS/LAPACK backend.
 
 ## License
 Dual-licensed to be compatible with the Rust project.
