+++ algorithms/linfa-pls/src/errors.rs
@@ -1,3 +1,6 @@
+#[cfg(not(feature = "blas"))]
+use linfa_linalg::LinalgError;
+#[cfg(feature = "blas")]
 use ndarray_linalg::error::LinalgError;
 use thiserror::Error;
 pub type Result<T> = std::result::Result<T, PlsError>;
