+++ algorithms/linfa-pls/src/pls_svd.rs
@@ -2,7 +2,10 @@ use crate::errors::{PlsError, Result};
 use crate::{utils, Float};
 use linfa::dataset::{WithLapack, WithoutLapack};
 use linfa::{dataset::Records, traits::Fit, traits::Transformer, DatasetBase};
+#[cfg(not(feature = "blas"))]
+use linfa_linalg::svd::*;
 use ndarray::{s, Array1, Array2, ArrayBase, Data, Ix2};
+#[cfg(feature = "blas")]
 use ndarray_linalg::svd::*;
 #[cfg(feature = "serde")]
 use serde_crate::{Deserialize, Serialize};
@@ -71,7 +74,11 @@ impl<F: Float, D: Data<Elem = F>> Fit<ArrayBase<D, Ix2>, ArrayBase<D, Ix2>, PlsE
 
         // Compute SVD of cross-covariance matrix
         let c = x.t().dot(&y);
-        let (u, _, vt) = c.with_lapack().svd(true, true)?;
+        let d = c.with_lapack().svd(true, true)?;
+        #[cfg(feature = "blas")]
+        let (u, _, vt) = d;
+        #[cfg(not(feature = "blas"))]
+        let (u, _, vt) = d.sort_svd_desc();
         // safe unwraps because both parameters are set to true in above call
         let u = u.unwrap().slice_move(s![.., ..self.n_components]);
         let vt = vt.unwrap().slice_move(s![..self.n_components, ..]);
