+++ algorithms/linfa-pls/src/utils.rs
@@ -2,9 +2,11 @@ use linfa::{
     dataset::{WithLapack, WithoutLapack},
     DatasetBase, Float,
 };
+#[cfg(not(feature = "blas"))]
+use linfa_linalg::svd::*;
 use ndarray::{s, Array1, Array2, ArrayBase, ArrayView2, Axis, Data, DataMut, Ix1, Ix2, Zip};
+#[cfg(feature = "blas")]
 use ndarray_linalg::svd::*;
-use ndarray_linalg::Scalar;
 use ndarray_stats::QuantileExt;
 
 pub fn outer<F: Float>(
@@ -21,7 +23,10 @@ pub fn outer<F: Float>(
 /// Calculates the pseudo inverse of a matrix
 pub fn pinv2<F: Float>(x: ArrayView2<F>, cond: Option<F>) -> Array2<F> {
     let x = x.with_lapack();
+    #[cfg(feature = "blas")]
     let (opt_u, s, opt_vh) = x.svd(true, true).unwrap();
+    #[cfg(not(feature = "blas"))]
+    let (opt_u, s, opt_vh) = x.svd(true, true).unwrap().sort_svd_desc();
     let u = opt_u.unwrap();
     let vh = opt_vh.unwrap();
 
@@ -38,11 +43,7 @@ pub fn pinv2<F: Float>(x: ArrayView2<F>, cond: Option<F>) -> Array2<F> {
     let mut ucut = u.slice_move(s![.., ..rank]);
     ucut /= &s.slice(s![..rank]).mapv(F::Lapack::cast);
 
-    vh.slice(s![..rank, ..])
-        .t()
-        .dot(&ucut.t())
-        .mapv(|v| v.conj())
-        .without_lapack()
+    vh.slice(s![..rank, ..]).t().dot(&ucut.t()).without_lapack()
 }
 
 #[allow(clippy::type_complexity)]
