+++ algorithms/linfa-preprocessing/Cargo.toml
@@ -13,13 +13,14 @@ readme = "README.md"
 keywords = ["machine-learning", "linfa", "ai", "ml", "preprocessing"]
 categories = ["algorithms", "mathematics", "science"]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+blas = ["ndarray-linalg", "linfa/ndarray-linalg"]
 
 [dependencies]
-
-linfa = { version = "0.5.0", path = "../..", features = ["ndarray-linalg"] }
-ndarray = { version = "0.15", features = ["approx", "blas"] }
-ndarray-linalg = { version = "0.14" }
+linfa = { version = "0.5.0", path = "../.." }
+ndarray = { version = "0.15", features = ["approx"] }
+ndarray-linalg = { version = "0.14", optional = true }
+linfa-linalg = { version = "0.1", default-features = false }
 ndarray-stats = "0.5"
 thiserror = "1.0"
 approx = { version = "0.4" }
