+++ algorithms/linfa-reduction/README.md
@@ -20,3 +20,12 @@ There is an usage example in the `examples/` directory. To run, use:
 $ cargo run --release --example diffusion_map
 $ cargo run --release --example pca
 ```
+
+## BLAS/LAPACK backend
+
+See [this section](../../README.md#blaslapack-backend) to enable an external BLAS/LAPACK backend.
+
+## License
+Dual-licensed to be compatible with the Rust project.
+
+Licensed under the Apache License, Version 2.0 <http://www.apache.org/licenses/LICENSE-2.0> or the MIT license <http://opensource.org/licenses/MIT>, at your option. This file may not be copied, modified, or distributed except according to those terms.
