+++ algorithms/linfa-reduction/src/diffusion_map/algorithms.rs
@@ -6,13 +6,20 @@
 //! diffusion graph with the (i, j) entry the probability that a diffusion happens from point i to
 //! j.
 //!
+#[cfg(not(feature = "blas"))]
+use linfa_linalg::{
+    eigh::*,
+    lobpcg::{self, LobpcgResult, Order as TruncatedOrder},
+};
 use ndarray::{Array1, Array2};
+#[cfg(feature = "blas")]
 use ndarray_linalg::{eigh::EighInto, lobpcg, lobpcg::LobpcgResult, Scalar, TruncatedOrder, UPLO};
 use ndarray_rand::{rand_distr::Uniform, RandomExt};
 
 use linfa::dataset::{WithLapack, WithoutLapack};
 use linfa::{traits::Transformer, Float};
 use linfa_kernel::Kernel;
+use rand::{prelude::SmallRng, SeedableRng};
 
 use super::hyperparams::DiffusionMapValidParams;
 
@@ -96,6 +103,7 @@ impl<F: Float> DiffusionMap<F> {
     }
 }
 
+#[allow(unused)]
 fn compute_diffusion_map<F: Float>(
     kernel: &Kernel<F>,
     steps: usize,
@@ -106,10 +114,9 @@ fn compute_diffusion_map<F: Float>(
     assert!(embedding_size < kernel.size());
 
     let d = kernel.sum().mapv(|x| x.recip());
-    let d2 = d.mapv(|x| x.powf(F::cast(0.5 + alpha)));
 
-    // use full eigenvalue decomposition for small problem sizes
     let (vals, vecs) = if kernel.size() < 5 * embedding_size + 1 {
+        // use full eigenvalue decomposition for small problem sizes
         let mut matrix = kernel.dot(&Array2::from_diag(&d).view());
         matrix
             .columns_mut()
@@ -118,20 +125,30 @@ fn compute_diffusion_map<F: Float>(
             .for_each(|(mut a, b)| a *= *b);
 
         let matrix = matrix.with_lapack();
-        let (vals, vecs) = matrix.eigh_into(UPLO::Lower).unwrap();
-        let (vals, vecs) = (vals.slice_move(s![..; -1]), vecs.slice_move(s![.., ..; -1]));
+        // Calculate the eigen decomposition sorted from largest to lowest
+        #[cfg(feature = "blas")]
+        let (vals, vecs) = {
+            let (vals, vecs) = matrix.eigh_into(UPLO::Lower).unwrap();
+            (
+                vals.slice_move(s![..; -1]).mapv(Scalar::from_real),
+                vecs.slice_move(s![.., ..; -1]),
+            )
+        };
+        #[cfg(not(feature = "blas"))]
+        let (vals, vecs) = matrix.eigh_into().unwrap().sort_eig_desc();
         (
-            vals.slice_move(s![1..=embedding_size])
-                .mapv(Scalar::from_real),
+            vals.slice_move(s![1..=embedding_size]),
             vecs.slice_move(s![.., 1..=embedding_size]),
         )
     } else {
+        let d2 = d.mapv(|x| x.powf(F::cast(0.5 + alpha)));
         // calculate truncated eigenvalue decomposition
         let x = guess
             .unwrap_or_else(|| {
-                Array2::random(
+                Array2::random_using(
                     (kernel.size(), embedding_size + 1),
                     Uniform::new(0.0f64, 1.0),
+                    &mut SmallRng::seed_from_u64(31),
                 )
                 .mapv(F::cast)
             })
@@ -156,13 +173,18 @@ fn compute_diffusion_map<F: Float>(
             x,
             |_| {},
             None,
-            1e-15,
+            1e-7,
             200,
             TruncatedOrder::Largest,
         );
 
         let (vals, vecs) = match result {
+            #[cfg(feature = "blas")]
             LobpcgResult::Ok(vals, vecs, _) | LobpcgResult::Err(vals, vecs, _, _) => (vals, vecs),
+            #[cfg(not(feature = "blas"))]
+            LobpcgResult::Ok(lobpcg) | LobpcgResult::Err((_, Some(lobpcg))) => {
+                (lobpcg.eigvals, lobpcg.eigvecs)
+            }
             _ => panic!("Eigendecomposition failed!"),
         };
