+++ algorithms/linfa-reduction/src/error.rs
@@ -1,7 +1,9 @@
 use thiserror::Error;
+
 pub type Result<T> = std::result::Result<T, ReductionError>;
 
 #[derive(Error, Debug)]
+#[non_exhaustive]
 pub enum ReductionError {
     #[error("At least 1 sample needed")]
     NotEnoughSamples,
@@ -9,8 +11,11 @@ pub enum ReductionError {
     EmbeddingTooSmall(usize),
     #[error("Number of steps zero in diffusion map operator")]
     StepsZero,
+    #[cfg(feature = "blas")]
     #[error(transparent)]
-    LinalgError(#[from] ndarray_linalg::error::LinalgError),
+    LinalgBlasError(#[from] ndarray_linalg::error::LinalgError),
+    #[error(transparent)]
+    LinalgError(#[from] linfa_linalg::LinalgError),
     #[error(transparent)]
     LinfaError(#[from] linfa::error::Error),
 }
