+++ algorithms/linfa-tsne/README.md
@@ -17,10 +17,10 @@ It wraps the [bhtsne](https://github.com/frjnn/bhtsne) crate, all kudos to them.
 
 ## Examples
 
-There is an usage example in the `examples/` directory. The example uses a BLAS backend, to run it and use the `intel-mkl` library do:
+There is an usage example in the `examples/` directory. To run it, do:
 
 ```bash
-$ cargo run --example tsne --features linfa/intel-mkl-system
+$ cargo run --example tsne
 ```
 
 You have to install the `gnuplot` library for plotting. Also take a look at the [README](https://github.com/rust-ml/linfa#blaslapack-backend) to see what BLAS/LAPACK backends are possible.
