+++ algorithms/linfa-tsne/examples/iris.dat
@@ -1,0 +1,150 @@
+-52.352851291391396 21.265257460427318 0
+-40.19390071862327 25.42788468872362 0
+-62.94193180010778 17.976110447403727 0
+-66.42886081504194 17.575522637113448 0
+-52.68607356903733 16.060805601264946 0
+-49.85256456053993 7.310720308312396 0
+-60.79960475942025 12.12748614304463 0
+-50.0579072709762 20.155979924788685 0
+-69.83044236029647 16.567545010548237 0
+-42.207261761145745 25.210411636680885 0
+-55.54435866939564 26.656612965791847 0
+-59.010870296631445 16.5251933324741 0
+-39.24858893051477 27.02132653969581 0
+-69.20716250078699 14.701455430201515 0
+-57.59823791779487 29.965482684295427 0
+-44.94602355723804 7.438717198198474 0
+-48.580962143960726 6.145869738250896 0
+-52.70863455046094 18.83842238946077 0
+-56.41428395325541 28.63678851321139 0
+-51.457726308011246 9.819944763993027 0
+-50.98779064540573 27.697376494148536 0
+-53.20477379599081 10.230375583022697 0
+-58.87555932220622 10.697540232115447 0
+-56.95625269166942 18.058389372135167 0
+-60.35383330333446 15.555511473181198 0
+-40.72160914286805 23.457086326163832 0
+-56.77530189071385 16.174973899927586 0
+-52.28059008283051 23.60938248354346 0
+-50.34510414530942 24.993186393680347 0
+-64.5605270567803 18.33181239169853 0
+-42.58592957477592 28.42727288251223 0
+-52.13989305755905 25.899734416606506 0
+-48.13822209817624 8.060474001686558 0
+-46.381260981417064 6.370287378194083 0
+-38.01548337118188 25.331080452519274 0
+-45.109840508618774 23.07934363064878 0
+-52.857545721459346 28.79388573657378 0
+-38.772467453251956 23.613624445302296 0
+-68.009619981141 15.914457039020554 0
+-49.92221592134641 22.423032776677974 0
+-53.98956378683552 17.11465806141574 0
+6.254251161376725 8.337098933810918 0
+-65.92731800021917 14.072020209526485 0
+-57.13545268826184 10.40530730936478 0
+-54.633048163285196 8.845002721199025 0
+-41.104897093488695 28.63893573611483 0
+-52.26420945098772 8.381284827213165 0
+-64.59838356863162 16.353154669352723 0
+-55.64819746235811 25.127673289208282 0
+-47.598593738385695 22.019984940034128 0
+40.60906342548215 7.243872519136638 1
+14.311284177484078 -25.40049268159925 1
+39.675013303212566 5.521486933102545 1
+11.015891249180223 5.13252860466794 1
+31.652531996088257 6.187000816674536 1
+15.133954442085592 -5.129484459839747 1
+20.080420418630172 -28.019388983929254 1
+6.899161010790488 2.7585553920798898 1
+35.34201498615892 7.936301129386888 1
+41.590606008130074 -31.915483075517514 1
+7.7368503367150545 6.4901883826507 1
+18.85590437734602 -11.929469445011511 1
+19.010963664816547 15.465375896006227 1
+23.390713990494472 -5.1881178307558455 1
+15.270220466604933 -10.518435347230481 1
+37.92406766984123 5.893602634285264 1
+36.221214079679676 -34.424515797634506 1
+19.189299710108145 8.728027046355184 1
+19.596007660321177 13.307093520588827 1
+14.458569685162209 7.839486134976738 1
+29.92286145394733 -33.407877234177164 1
+26.880342924353165 5.4831163412108825 1
+22.593551049781535 9.176930208531264 1
+24.575960952091947 5.543545445708245 1
+33.40237071052989 6.27800514988937 1
+35.945052965734355 5.510537181608577 1
+36.684255059046684 9.398625893754993 1
+37.58426867934996 1.6367232375523413 1
+20.30931726734965 -8.617715814764683 1
+17.262631606233303 9.691443059437322 1
+12.432618695081072 6.752188028586107 1
+12.606259987729917 8.471488775001461 1
+19.363072235038256 6.706616009392266 1
+17.296712471805925 -3.6084015560489315 1
+36.30466329063912 -32.280010811660404 1
+27.952359948349237 -32.211219480828866 1
+37.619338532932595 3.70149265442679 1
+20.42018431022558 14.840139565065142 1
+16.633302192995274 -12.76374295023074 1
+10.749935219913796 0.5192644170316616 1
+11.301066964103176 -0.986202851235202 1
+23.0381537361522 -7.2921835439938505 1
+17.726674899411822 7.554462775447745 1
+7.758920572794544 4.3316387521803925 1
+13.211353271739013 -4.038455387313851 1
+17.481802988608955 -10.047225622921077 1
+16.14555305740276 -8.561866239795036 1
+28.574910477776204 5.2232713967033675 1
+8.267131202978755 1.5415449003058548 1
+14.468631468866068 -6.96346201999712 1
+28.75641139221956 -39.38072543192582 2
+39.604935518493406 -35.04544385002682 2
+40.39905228042734 0.11457661238601347 2
+22.888491997047886 -17.90354011482502 2
+24.90338421488204 -24.569779702120655 2
+45.68514579728229 3.8324106560141113 2
+39.83626553839454 -29.239967142618834 2
+46.09056478987343 5.96826767811466 2
+24.139340915106164 9.63380652044945 2
+21.52685853976281 -35.575882415755636 2
+22.28149978421835 -28.766605678595276 2
+21.094650395216025 -3.6246010965878837 2
+17.596242695172794 -24.696744127511757 2
+41.0161536685218 -33.752257070962955 2
+35.782325616899556 -29.43436702283855 2
+27.41046448589042 -34.4583021053664 2
+19.34209273965296 -22.789793622089004 2
+45.16719385537672 15.85120421646311 2
+48.380410853092926 6.019588059503489 2
+14.824796121927633 5.163680208112121 2
+20.36118150887557 -31.32044921813969 2
+37.15491181331178 -30.56086157459833 2
+48.080738801142346 7.880054166272845 2
+22.698962668999986 -3.107627652859589 2
+22.401317121836772 -31.038076424454676 2
+42.42371999021448 1.9334841354607413 2
+20.63535696358845 -7.080227579062087 2
+21.37789234274928 -15.105374473891573 2
+23.79431004210474 -19.174233762202622 2
+44.313720733417654 6.882909477940987 2
+46.08623239473761 7.696702316965035 2
+45.00828446287442 14.595086001331971 2
+25.312952149246858 -18.73682077793249 2
+23.657389103215063 -0.39387998497938176 2
+20.68853085361109 2.4964727015542647 2
+44.005491706805905 3.6888717808677725 2
+27.123837852867922 -38.559113789610926 2
+22.41759768786052 -24.174823942722433 2
+17.307971266369687 -15.711798721422797 2
+16.5252983708557 -26.38498626888663 2
+24.252164062082215 -32.07498430651642 2
+17.564738893370667 -28.162050034491898 2
+39.245931178174125 -33.08898021959067 2
+22.332565555800016 -32.70695729036592 2
+25.609033239142253 -35.358112482599935 2
+21.887853629376917 -26.60668397203725 2
+20.965216097253343 -1.34049022032243 2
+20.93170652919792 -24.215825788747626 2
+28.54012284344299 -37.58544550692399 2
+33.79915935846433 -33.033588345593266 2
